+++ CHANGELOG.md
@@ -83,3 +83,10 @@
 - Readme
 - Code cleanup
 - Unity 2019 compatibility (can't gaurantee it works, I wont test it, but should at least compile)
+
+## [0.1.9] - 2024-2-25
+
+### Added
+
+- Function to change which folder generated assets are saved to
+- Ability to edit step delay (not recommended)